Created data loader for licenses
