+++ docs/.vitepress/theme/data/licenses/validateLicense.ts
